refactor(mesh): extract pure Kotlin crypto to commonMain
- Add PureKotlinCrypto with shared AES-128 block cipher and SHA-256
- All hex constants converted to decimal for KMP compatibility
  (hex > 0x7FFFFFFF parsed as Long in Kotlin common code)
- iOS CryptoEngine now delegates to PureKotlinCrypto instead of
  duplicating the AES + SHA-256 implementations
- Single source of truth for crypto primitives across all platforms
